fix(server): fix slop abstractions
