+++ forge-engine/crates/forge-server/src/connection.rs
@@ -7,9 +7,10 @@ use tokio::sync::mpsc;
 use tokio_tungstenite::tungstenite::Message;
 use tracing::{debug, info, warn};
 
+use crate::cleanup::mark_disconnected;
 use crate::error::ServerError;
 use crate::lobby;
-use crate::protocol::{ClientMessage, RoomStatus, ServerMessage};
+use crate::protocol::{ClientMessage, ServerMessage};
 use crate::state::{ConnectedPlayer, ServerState};
 
 type WsSender = futures_util::stream::SplitSink<
@@ -828,173 +829,12 @@ fn handle_client_message(
     }
 }
 
-pub(crate) fn mark_disconnected(state: &Arc<ServerState>, player_id: &str, our_generation: u64) {
-    let (username, room_id) = {
-        if let Some(mut player) = state.players.get_mut(player_id) {
-            if player.generation != our_generation {
-                info!(
-                    "[disconnect] '{}' old connection cleaned up (session reclaimed by new connection)",
-                    player.username
-                );
-                return;
-            }
-            player.connected = false;
-            player.disconnected_at = Some(Instant::now());
-            (player.username.clone(), player.room_id.clone())
-        } else {
-            return;
-        }
-    };
-
-    if let Some(rid) = &room_id {
-        let room_status = state.rooms.get(rid).map(|r| r.status.clone());
-
-        match room_status {
-            Some(RoomStatus::InGame) => {
-                // InGame: always preserve session for reconnection.
-                let host_without_player = state
-                    .rooms
-                    .get(rid)
-                    .map(|room| room.is_host(player_id) && !room.host_is_player())
-                    .unwrap_or(false);
-                if host_without_player {
-                    info!(
-                        "[cleanup] hosted in-game room {} lost its non-playing host -- removing",
-                        &rid[..8]
-                    );
-                    remove_room_and_clear_sessions(state, rid);
-                    return;
-                }
-
-                if let Some(mut room) = state.rooms.get_mut(rid) {
-                    room.set_connected(player_id, false);
-                } else {
-                    return;
-                }
-
-                info!(
-                    "[disconnect] '{}' marked disconnected in in-game room {} (session preserved)",
-                    username,
-                    &rid[..8]
-                );
-                broadcast_to_room(
-                    state,
-                    rid,
-                    &ServerMessage::PlayerDisconnected {
-                        username: username.clone(),
-                    },
-                );
-
-                if let Some(room) = state.rooms.get(rid) {
-                    broadcast_to_room(
-                        state,
-                        rid,
-                        &ServerMessage::RoomUpdate {
-                            room: room.to_room_info(),
-                        },
-                    );
-                }
-            }
-            Some(RoomStatus::Lobby) => {
-                // Lobby: treat like a leave — remove player, clean up room, free username
-                info!(
-                    "[disconnect] '{}' disconnected from lobby room {} -- treating as leave",
-                    username,
-                    &rid[..8]
-                );
-
-                let remove_hosted_room = state
-                    .rooms
-                    .get(rid)
-                    .map(|room| room.is_host(player_id) && !room.host_is_player())
-                    .unwrap_or(false);
-                if remove_hosted_room {
-                    info!(
-                        "[cleanup] hosted lobby room {} lost its non-playing host -- removing",
-                        &rid[..8]
-                    );
-                    remove_room_and_clear_sessions(state, rid);
-                    return;
-                }
-
-                let room_empty = {
-                    if let Some(mut room) = state.rooms.get_mut(rid) {
-                        room.remove_participant(player_id);
-                        room.is_empty()
-                    } else {
-                        false
-                    }
-                };
-
-                if let Some(mut player) = state.players.get_mut(player_id) {
-                    player.room_id = None;
-                }
-
-                if room_empty {
-                    info!(
-                        "[cleanup] lobby room {} is now empty -- removing",
-                        &rid[..8]
-                    );
-                    remove_room_and_clear_sessions(state, rid);
-                } else {
-                    broadcast_to_room(
-                        state,
-                        rid,
-                        &ServerMessage::PlayerLeft {
-                            room_id: rid.clone(),
-                            username: username.clone(),
-                        },
-                    );
-                    if let Some(room) = state.rooms.get(rid) {
-                        broadcast_to_room(
-                            state,
-                            rid,
-                            &ServerMessage::RoomUpdate {
-                                room: room.to_room_info(),
-                            },
-                        );
-                    }
-                }
-
-                info!("[cleanup] '{}' removed (disconnected from lobby)", username);
-                state.players.remove(player_id);
-            }
-            None => {
-                info!("[cleanup] '{}' removed (room no longer exists)", username);
-                state.players.remove(player_id);
-            }
-        }
-    } else {
-        info!("[cleanup] '{}' removed (was not in a room)", username);
-        state.players.remove(player_id);
-    }
-}
-
 fn get_username(state: &Arc<ServerState>, player_id: &str) -> String {
     state
         .players
         .get(player_id)
         .map(|p| p.username.clone())
         .unwrap_or_default()
-}
-
-pub(crate) fn remove_room_and_clear_sessions(state: &Arc<ServerState>, room_id: &str) {
-    state.rooms.remove(room_id);
-    let player_ids = state
-        .players
-        .iter()
-        .filter_map(|entry| {
-            entry
-                .value()
-                .room_id
-                .as_deref()
-                .is_some_and(|rid| rid == room_id)
-                .then(|| entry.key().clone())
-        })
-        .collect::<Vec<_>>();
-    for player_id in player_ids {
-        state.players.remove(&player_id);
-    }
 }
 
 fn msg_type_of(msg: &ServerMessage) -> &'static str {
