+++ forge-engine/crates/forge-server/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cleanup;
 pub mod config;
 pub mod connection;
 pub mod error;
