+++ forge-engine/crates/forge-server/src/main.rs
@@ -1,6 +1,7 @@
 use std::net::SocketAddr;
 use std::sync::Arc;
 
+mod cleanup;
 mod config;
 mod connection;
 mod error;
